// src/shared/dataprovider/api/model/transactionSummaryDataTotal.ts
/**
 * DynastyGateway API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface TransactionSummaryDataTotal { 
    TotalCashin?: number;
    TotalCashout?: number;
    TotalWithdrawal?: number;
    TotalSuccessfulTransaction?: number;
    TotalClosedTransaction? : number;
    TotalCancelledTransaction?: number;
    NetTotalCashin? : number;
    NetTotalCashout? : number;
    NetTotalWithdrawal? : number;
    TotalCashinFees? : number;
    TotalCashoutFees? : number;
    TotalFailedTransaction? : number;
    TotalWithdrawalFees? : number;
}